Popover down does not fit when using 13" screens
Currently the configuration of the bootstrap popover has the placement set to `bottom`. This means that smaller screens (for example a 13" MBP) would not be able to see the full content of the popover. By changing this to `right`, it would allow the content to be displayed correctly, without going outside the screen!

Discovered this problem today in an instance of avrodoc, and found it already reported at https://github.com/ga4gh/dwg-website/issues/4